requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
akismet==1.2.1
alabaster==0.7.12
amqp==5.1.1
analytics-python==1.4.0
appnope==0.1.3
appoptics-metrics==5.1.0
asgiref==3.5.2
astroid==2.11.5
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
awesome-slugify==1.6.5
Babel==2.10.1
backcall==0.2.0
backoff==1.10.0
backports.csv==1.0.7
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
boto3==1.24.2
botocore==1.27.2
cached-property==1.5.2
cattrs==22.1.0
cchardet==2.1.7
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confusable-homoglyphs==3.2.0
cryptography==37.0.2
curlify==2.2.1
dataclasses-json==0.5.7
datapackage==1.15.2
datetime-truncate==1.1.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dict2xml==1.7.1
diff-match-patch==20200713
dill==0.3.5.1
Django==3.2.13
django-activity-stream==1.4.1
django-admin-charts==1.0.0
django-admin-tools==0.9.2
django-allauth==0.50.0
django-appconf==1.0.5
django-avatar==5.0.0
django-axes==5.35.0
django-bootstrap-form==3.4
django-bootstrap4-form==4.0.1
django-bower==5.2.0
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-cors-headers==3.12.0
django-cors-middleware==1.5.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
-e git+https://github.com/dimagi/django-digest.git@6bf61ec08502fd3545d4f2c0838b6cb15e7ffa92#egg=django_digest
django-easy-pjax==1.3.0
django-extensions==3.1.5
django-filter==21.1
django-formtools==2.3
django-genericadmin==0.7.0
django-guardian==2.4.0
django-helpdesk==0.3.4
django-hijack==3.2.1
django-import-export==2.8.0
django-invitations==1.9.3
django-ipware==4.0.2
django-jsonfield==1.4.1
django-jsonview==2.0.0
django-mailer==2.2
django-markdown-deux==1.0.5
django-mathfilters==1.0.0
django-model-utils==4.2.0
django-multidb-router==0.10
django-multiselectfield==0.1.12
django-next-url-mixin==0.3.0
django-nose==1.4.7
django-notifications-hq==1.7.0
django-nvd3==0.9.7
django-oauth-toolkit==2.0.0
django-oidc-provider==0.7.0
django-ordered-model==3.6
django-organizations==2.0.2
django-otp==1.1.3
django-phonenumber-field==6.1.0
django-qsstats-magic==1.1.0
django-query-builder==2.0.1
django-recaptcha3==0.4.0
django-registration==3.3
django-registration-redux==2.10
django-render-block==0.9.1
django-rest-knox==4.2.0
django-reversion==5.0.0
django-selectable==1.3.0
django-sequences==2.7
django-sql-explorer==2.4.1
django-tables2==2.4.1
django-tables2-reports==0.1.3
django-taggit==3.0.0
django-tellme==0.7.3
django-templated-email==3.0.0
django-timezone-field==5.0
django-tinymce==3.4.0
django-two-factor-auth==1.13.2
django-user-accounts==3.2.0
django-user-sessions==1.7.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-gis==1.0
djangorestframework-guardian==0.3.0
djangorestframework-jsonapi==5.0.0
djangorestframework-jsonp==1.0.2
djangorestframework-xml==2.0.0
djcacheutils==3.0.0
docutils==0.18.1
dpath==2.0.6
elaphe3==0.2.0
email-reply-parser==0.5.12
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc7
executing==0.8.3
flake8==4.0.1
fleming==0.7.0
flower==1.0.0
FormEncode==2.0.1
frozenlist==1.3.0
future==0.18.2
geoip2==4.5.0
geojson==2.5.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
httmock==1.4.0
httplib2==0.20.4
humanize==4.1.0
idna==3.3
idna-ssl==1.1.0
ijson==3.1.4
imagesize==1.3.0
importlib-metadata==4.11.4
importlib-resources==5.7.1
inflection==0.5.1
ipdb==0.13.9
ipython==8.4.0
ipython-genutils==0.2.0
isodate==0.6.1
isort==5.10.1
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
jsonfield==3.1.0
jsonlines==3.0.0
jsonpickle==2.2.0
jsonpointer==2.3
jsonschema==4.6.0
jwcrypto==1.3.1
kombu==5.2.4
lazy-object-proxy==1.7.1
linear-tsv==1.1.0
linecache2==1.0.0
lockfile==0.12.2
lxml==4.9.0
Markdown==3.3.7
markdown2==2.4.3
MarkupPy==1.14
MarkupSafe==2.1.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
mccabe==0.6.1
mock==4.0.3
modilabs-python-utils==0.1.5
monotonic==1.6
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
nameparser==1.1.1
nose==1.3.7
numpy==1.22.4
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.49
pickleshare==0.7.5
pika==1.3.2
Pillow==9.1.1
pinax-invitations==7.0.0
pinax-teams==1.0.5
pinax-templates==2.0.6
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.21.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
-e git+https://github.com/Kirembu/py-asterisk.git@195483d5382c8bef9b04b2d5745fd852336089e9#egg=py_Asterisk
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.14.1
pyflakes==2.4.0
Pygments==2.12.0
pyjwkest==1.4.2
PyJWT==2.4.0
pylibmc==1.6.1
pylint==2.14.0
pylint-django==2.5.3
pylint-plugin-utils==0.7
pymongo==4.1.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-digest==1.7
python-facebook-api==0.14.0
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
python-stdnum==1.17
python3-openid==3.2.0
pytz==2022.1
pyxform==1.10.0
PyYAML==6.0
qrcode==7.3.1
raven==6.10.0
recaptcha-client==1.0.6
redis==4.3.3
regex==2022.6.2
requests==2.27.1
requests-mock==1.9.3
requests-oauthlib==1.3.1
responses==0.21.0
rfc3986==2.0.0
rsa==4.8
s3transfer==0.6.0
savReaderWriter==3.4.2
sentry-sdk==1.5.12
simplejson==3.17.6
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.37
sqlparse==0.4.2
stack-data==0.2.0
suds==1.1.1
swapper==1.3.0
tableschema==1.20.2
tablib==3.2.1
tabulator==1.53.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tornado==6.1
traceback2==1.4.0
traitlets==5.2.2.post1
tweepy==4.10.0
twilio==7.9.2
typed-ast==1.5.4
typing-inspect==0.7.1
typing_extensions==4.2.0
tzdata==2022.1
unicode-slugify==0.1.5
unicodecsv==0.14.1
Unidecode==0.4.21
unittest2==1.1.0
urllib3==1.26.9
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
xmltodict==0.13.0
yapf==0.32.0
yarl==1.7.2
YubiOTP==1.0.0
-e git+https://github.com/Kirembu/zerxis-invitations.git@41c5dcfa090f91d8846a203d9080af67d12cfb66#egg=zerxis_invitations
-e git+https://github.com/Kirembu/zerxis-user-accounts.git@42338ed98d0249335417ece948d7ea4271a66180#egg=zerxis_user_accounts
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
